modified:   index.html 	modified:   jsTest.html Nothing done to modified files
	new file:   requirements.xml
	new file:   rtm.html
Requirements.xml is a sample requirements doc to be used in the initial creation of the RTM.
rtm.html is the html page we will use to develop our RTM. So far it takes the requirements and puts them in an HTML table.
ToDo: create use and test case XML's
Compare these to requirements and mark table
Export html table as xsl table.