Clicking first section fixed
Had to move the section info to the TOC div in XSL. The show section JS
was overwriting the first section div each time. I changed it so the
showSection JS wrote to the view div. This may have ill effects on some
edit functionality but I'll address that as I find it.