TestCase fixing
Changes test case back to NotionalXSL so it works with our app. Also
fixed one thing where if you used the add section or requirement it
wasn't adding the index to the paras. I think this is one of the
problems Michael was talking about however its not completely fixed yet.